Fix admin user createion #191
The form wasn't wrapping the submit buttons so the actions to save were never firing
